requirements.txt
streamlit==1.40.2
langchain-openai==0.2.14
langchain-core==0.3.29
python-dotenv==1.0.1